+++ api/src/service/user_permission_revoke.ts
@@ -1,35 +1,33 @@
 import Intent from "../authz/intents";
 import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
-
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
 import { Identity } from "./domain/organization/identity";
 import { ServiceUser } from "./domain/organization/service_user";
-import * as Project from "./domain/workflow/project";
-import * as ProjectPermissionGrant from "./domain/workflow/project_permission_grant";
+import * as UserPermissionRevoke from "./domain/organization/user_permission_revoke";
+import * as UserRecord from "./domain/organization/user_record";
 import { store } from "./store";
+import * as UserQuery from "./user_query";
 
 export { RequestData } from "./domain/workflow/project_create";
 
-export async function grantProjectPermission(
+export async function revokeUserPermission(
   conn: ConnToken,
   ctx: Ctx,
   serviceUser: ServiceUser,
-  projectId: Project.Id,
-  grantee: Identity,
+  userId: UserRecord.Id,
+  revokee: Identity,
   intent: Intent,
 ): Promise<void> {
   const result = await Cache.withCache(conn, ctx, async cache =>
-    ProjectPermissionGrant.grantProjectPermission(ctx, serviceUser, projectId, grantee, intent, {
-      getProject: async id => {
-        return cache.getProject(id);
-      },
+    UserPermissionRevoke.revokeUserPermission(ctx, serviceUser, userId, revokee, intent, {
+      getTargetUser: id => UserQuery.getUser(conn, ctx, serviceUser, id),
     }),
   );
   if (Result.isErr(result)) return Promise.reject(result);
 
-  for (const event of result.newEvents) {
+  for (const event of result) {
     await store(conn, ctx, event);
   }
 }